fix: complete network pattern matching and enhance production readiness
- Fix pattern matching exhaustiveness errors in peer_manager.rs
- Resolve ExtensionError::IoError type mismatches in plugins
- Fix Cow<str> comparison issues in tokens_tracker.rs
- Add P2PConfig imports and fix SocketAddr defaults
- Enhanced production-readiness.sh with Rust-specific patterns
- Remove missing example reference from network Cargo.toml
- All compilation errors resolved and tests passing

+++ crates/plugins/src/oracle_service.rs
@@ -164,7 +164,7 @@ impl OracleServicePlugin {
         let client = Client::builder()
             .timeout(Duration::from_secs(self.http_config.timeout_seconds))
             .redirect(reqwest::redirect::Policy::limited(
-                self.http_config.max_redirects,
+                self.http_config.max_redirects as usize,
             ))
             .user_agent(&self.http_config.user_agent)
             .build()
